Move portions of CLI spec to new spec files
cli_spec.rb was very, very big, and it was hard to find things. Hopefully,
splitting off some sections into independent files will help.
