+++ spec/rubocop/cli_spec.rb
@@ -28,2349 +28,6 @@ describe RuboCop::CLI, :isolated_environment do
     File.expand_path(path)
   end
 
-  describe 'option' do
-    describe '--list-target-files' do
-      context 'when there are no files' do
-        it 'prints nothing with -L' do
-          cli.run ['-L']
-          expect($stdout.string).to be_empty
-        end
-
-        it 'prints nothing with --list-target-files' do
-          cli.run ['--list-target-files']
-          expect($stdout.string).to be_empty
-        end
-      end
-
-      context 'when there are some files' do
-        before(:each) do
-          create_file('show.rabl', 'object @user => :person')
-          create_file('app.rb', 'puts "hello world"')
-          create_file('Gemfile', ['source "https://rubygems.org"',
-                                  'gem "rubocop"'])
-          create_file('lib/helper.rb', 'puts "helpful"')
-        end
-
-        context 'when there are no includes or excludes' do
-          it 'prints known ruby files' do
-            cli.run ['-L']
-            expect($stdout.string.split("\n")).to match_array ['app.rb',
-                                                               'Gemfile',
-                                                               'lib/helper.rb']
-          end
-        end
-
-        context 'when there is an include and exclude' do
-          before(:each) do
-            create_file('.rubocop.yml', ['AllCops:',
-                                         '  Exclude:',
-                                         '    - Gemfile',
-                                         '  Include:',
-                                         '    - "**/*.rabl"'])
-          end
-
-          it 'prints the included files and not the excluded ones' do
-            cli.run ['--list-target-files']
-            expect($stdout.string.split("\n")).to match_array ['app.rb',
-                                                               'lib/helper.rb',
-                                                               'show.rabl']
-          end
-        end
-      end
-    end
-
-    describe '--version' do
-      it 'exits cleanly' do
-        expect(cli.run(['-v'])).to eq(0)
-        expect(cli.run(['--version'])).to eq(0)
-        expect($stdout.string).to eq((RuboCop::Version::STRING + "\n") * 2)
-      end
-    end
-
-    describe '--auto-correct' do
-      it 'does not correct ExtraSpacing in a hash that would be changed back' do
-        create_file('.rubocop.yml', ['Style/AlignHash:',
-                                     '  EnforcedColonStyle: table'])
-        source = ['hash = {',
-                  '  alice: {',
-                  '    age:  23,',
-                  "    role: 'Director'",
-                  '  },',
-                  '  bob:   {',
-                  '    age:  25,',
-                  "    role: 'Consultant'",
-                  '  }',
-                  '}']
-        create_file('example.rb', source)
-        expect(cli.run(['--auto-correct'])).to eq(1)
-        expect(IO.read('example.rb')).to eq(source.join("\n") + "\n")
-      end
-
-      it 'does not correct SpaceAroundOperators in a hash that would be ' \
-         'changed back' do
-        create_file('.rubocop.yml', ['Style/HashSyntax:',
-                                     '  EnforcedStyle: hash_rockets',
-                                     '',
-                                     'Style/AlignHash:',
-                                     '  EnforcedHashRocketStyle: table'])
-        source = ['a = { 1=>2, a => b }',
-                  'hash = {',
-                  '  :alice => {',
-                  '    :age  => 23,',
-                  "    :role => 'Director'",
-                  '  },',
-                  '  :bob   => {',
-                  '    :age  => 25,',
-                  "    :role => 'Consultant'",
-                  '  }',
-                  '}']
-        create_file('example.rb', source)
-        expect(cli.run(['--auto-correct'])).to eq(1)
-
-        # 1=>2 is changed to 1 => 2. The rest is unchanged.
-        # SpaceAroundOperators leaves it to AlignHash when the style is table.
-        expect(IO.read('example.rb')).to eq(['a = { 1 => 2, a => b }',
-                                             'hash = {',
-                                             '  :alice => {',
-                                             '    :age  => 23,',
-                                             "    :role => 'Director'",
-                                             '  },',
-                                             '  :bob   => {',
-                                             '    :age  => 25,',
-                                             "    :role => 'Consultant'",
-                                             '  }',
-                                             '}'].join("\n") + "\n")
-      end
-
-      it 'corrects IndentationWidth, RedundantBegin, and ' \
-         'RescueEnsureAlignment offenses' do
-        source = ['def verify_section',
-                  '      begin',
-                  '      scroll_down_until_element_exists',
-                  '      rescue',
-                  '        scroll_down_until_element_exists',
-                  '        end',
-                  'end']
-        create_file('example.rb', source)
-        expect(cli.run(['--auto-correct'])).to eq(0)
-        corrected = ['def verify_section',
-                     '  scroll_down_until_element_exists',
-                     'rescue',
-                     '  scroll_down_until_element_exists',
-                     'end',
-                     '']
-        expect(IO.read('example.rb')).to eq(corrected.join("\n"))
-      end
-
-      it 'corrects InitialIndentation offenses' do
-        source = ['  # comment 1',
-                  '',
-                  '  # comment 2',
-                  '  def func',
-                  '    begin',
-                  '      foo',
-                  '      bar',
-                  '    rescue',
-                  '      baz',
-                  '    end',
-                  '  end',
-                  ''].join("\n")
-        create_file('example.rb', source)
-        create_file('.rubocop.yml', ['Lint/DefEndAlignment:',
-                                     '  AutoCorrect: true'])
-        expect(cli.run(['--auto-correct'])).to eq(0)
-        corrected = ['# comment 1',
-                     '',
-                     '# comment 2',
-                     'def func',
-                     '  foo',
-                     '  bar',
-                     'rescue',
-                     '  baz',
-                     'end',
-                     '']
-        expect($stderr.string).to eq('')
-        expect(IO.read('example.rb')).to eq(corrected.join("\n"))
-      end
-
-      it 'corrects RedundantBegin offenses and fixes indentation etc' do
-        source = ['  def func',
-                  '    begin',
-                  '      foo',
-                  '      bar',
-                  '    rescue',
-                  '      baz',
-                  '    end',
-                  '  end',
-                  '',
-                  '  def func; begin; x; y; rescue; z end end',
-                  '',
-                  'def method',
-                  '  begin',
-                  '    BlockA do |strategy|',
-                  '      foo',
-                  '    end',
-                  '',
-                  '    BlockB do |portfolio|',
-                  '      foo',
-                  '    end',
-                  '',
-                  '  rescue => e # some problem',
-                  '    bar',
-                  '  end',
-                  'end',
-                  '',
-                  'def method',
-                  '  begin # comment 1',
-                  '    do_some_stuff',
-                  '  rescue # comment 2',
-                  '  end # comment 3',
-                  'end',
-                  ''].join("\n")
-        create_file('example.rb', source)
-        expect(cli.run(['--auto-correct'])).to eq(1)
-        corrected = ['def func',
-                     '  foo',
-                     '  bar',
-                     '  rescue',
-                     '    baz',
-                     '  end',
-                     '',
-                     'def func',
-                     '  x; y; rescue; z',
-                     'end',
-                     '',
-                     'def method',
-                     '  BlockA do |_strategy|',
-                     '    foo',
-                     '  end',
-                     '',
-                     '  BlockB do |_portfolio|',
-                     '    foo',
-                     '  end',
-                     '',
-                     'rescue => e # some problem',
-                     '  bar',
-                     'end',
-                     '',
-                     'def method',
-                     '  # comment 1',
-                     '  do_some_stuff',
-                     'rescue # comment 2',
-                     '  # comment 3',
-                     'end',
-                     '']
-        expect(IO.read('example.rb')).to eq(corrected.join("\n"))
-      end
-
-      it 'corrects Tab and IndentationConsistency offenses' do
-        source = ['  render_views',
-                  "    describe 'GET index' do",
-                  "\t    it 'returns http success' do",
-                  "\t    end",
-                  "\tdescribe 'admin user' do",
-                  '     before(:each) do',
-                  "\t    end",
-                  "\tend",
-                  '    end',
-                  '']
-        create_file('example.rb', source)
-        expect(cli.run(['--auto-correct'])).to eq(0)
-        corrected = ['render_views',
-                     "describe 'GET index' do",
-                     "  it 'returns http success' do",
-                     '  end',
-                     "  describe 'admin user' do",
-                     '    before(:each) do',
-                     '    end',
-                     '  end',
-                     'end',
-                     '']
-        expect(IO.read('example.rb')).to eq(corrected.join("\n"))
-      end
-
-      it 'corrects IndentationWidth and IndentationConsistency offenses' do
-        source = ["require 'spec_helper'",
-                  'describe ArticlesController do',
-                  '  render_views',
-                  '    describe "GET \'index\'" do',
-                  '            it "returns http success" do',
-                  '            end',
-                  '        describe "admin user" do',
-                  '             before(:each) do',
-                  '            end',
-                  '        end',
-                  '    end',
-                  'end']
-        create_file('example.rb', source)
-        expect(cli.run(['--auto-correct'])).to eq(0)
-        corrected = ["require 'spec_helper'",
-                     'describe ArticlesController do',
-                     '  render_views',
-                     "  describe \"GET 'index'\" do",
-                     "    it 'returns http success' do",
-                     '    end',
-                     "    describe 'admin user' do",
-                     '      before(:each) do',
-                     '      end',
-                     '    end',
-                     '  end',
-                     'end',
-                     '']
-        expect(IO.read('example.rb')).to eq(corrected.join("\n"))
-      end
-
-      it 'corrects SymbolProc and SpaceBeforeBlockBraces offenses' do
-        source = ['foo.map{ |a| a.nil? }']
-        create_file('example.rb', source)
-        expect(cli.run(['-D', '--auto-correct'])).to eq(0)
-        corrected = "foo.map(&:nil?)\n"
-        expect(IO.read('example.rb')).to eq(corrected)
-        uncorrected = $stdout.string.split($RS).select do |line|
-          line.include?('example.rb:') && !line.include?('[Corrected]')
-        end
-        expect(uncorrected).to be_empty # Hence exit code 0.
-      end
-
-      it 'corrects only IndentationWidth without crashing' do
-        source = ['foo = if bar',
-                  '  something',
-                  'elsif baz',
-                  '  other_thing',
-                  'else',
-                  '  fail',
-                  'end']
-        create_file('example.rb', source)
-        expect(cli.run(%w(--only IndentationWidth --auto-correct))).to eq(0)
-        corrected = ['foo = if bar',
-                     '        something',
-                     'elsif baz',
-                     '  other_thing',
-                     'else',
-                     '  fail',
-                     'end',
-                     ''].join("\n")
-        expect(IO.read('example.rb')).to eq(corrected)
-      end
-
-      it 'corrects complicated cases conservatively' do
-        # Two cops make corrections here; Style/BracesAroundHashParameters, and
-        # Style/AlignHash. Because they make minimal corrections relating only
-        # to their specific areas, and stay away from cleaning up extra
-        # whitespace in the process, the combined changes don't interfere with
-        # each other and the result is semantically the same as the starting
-        # point.
-        source = ['# encoding: utf-8',
-                  'expect(subject[:address]).to eq({',
-                  "  street1:     '1 Market',",
-                  "  street2:     '#200',",
-                  "  city:        'Some Town',",
-                  "  state:       'CA',",
-                  "  postal_code: '99999-1111'",
-                  '})']
-        create_file('example.rb', source)
-        expect(cli.run(['-D', '--auto-correct'])).to eq(0)
-        corrected =
-          ['# encoding: utf-8',
-           "expect(subject[:address]).to eq(street1:     '1 Market',",
-           "                                street2:     '#200',",
-           "                                city:        'Some Town',",
-           "                                state:       'CA',",
-           "                                postal_code: '99999-1111')"]
-        expect(IO.read('example.rb')).to eq(corrected.join("\n") + "\n")
-      end
-
-      it 'honors Exclude settings in individual cops' do
-        source = ['# encoding: utf-8',
-                  'puts %x(ls)']
-        create_file('example.rb', source)
-        create_file('.rubocop.yml', ['Style/CommandLiteral:',
-                                     '  Exclude:',
-                                     '    - example.rb'])
-        expect(cli.run(['--auto-correct'])).to eq(0)
-        expect($stdout.string).to include('no offenses detected')
-        expect(IO.read('example.rb')).to eq(source.join("\n") + "\n")
-      end
-
-      it 'corrects code with indentation problems' do
-        create_file('example.rb', ['# encoding: utf-8',
-                                   'module Bar',
-                                   'class Goo',
-                                   '  def something',
-                                   '    first call',
-                                   "      do_other 'things'",
-                                   '      if other > 34',
-                                   '        more_work',
-                                   '      end',
-                                   '  end',
-                                   'end',
-                                   'end',
-                                   '',
-                                   'module Foo',
-                                   'class Bar',
-                                   '',
-                                   '  stuff = [',
-                                   '            {',
-                                   "              some: 'hash',",
-                                   '            },',
-                                   '                 {',
-                                   "              another: 'hash',",
-                                   "              with: 'more'",
-                                   '            },',
-                                   '          ]',
-                                   'end',
-                                   'end'
-                                  ])
-        expect(cli.run(['--auto-correct'])).to eq(1)
-        expect(IO.read('example.rb'))
-          .to eq(['# encoding: utf-8',
-                  'module Bar',
-                  '  class Goo',
-                  '    def something',
-                  '      first call',
-                  "      do_other 'things'",
-                  '      more_work if other > 34',
-                  '    end',
-                  '  end',
-                  'end',
-                  '',
-                  'module Foo',
-                  '  class Bar',
-                  '    stuff = [',
-                  '      {',
-                  "        some: 'hash'",
-                  '      },',
-                  '      {',
-                  "        another: 'hash',",
-                  "        with: 'more'",
-                  '      }',
-                  '    ]',
-                  '  end',
-                  'end',
-                  ''].join("\n"))
-      end
-
-      it 'can change block comments and indent them' do
-        create_file('example.rb', ['# encoding: utf-8',
-                                   'module Foo',
-                                   'class Bar',
-                                   '=begin',
-                                   'This is a nice long',
-                                   'comment',
-                                   'which spans a few lines',
-                                   '=end',
-                                   '  def baz',
-                                   '    do_something',
-                                   '  end',
-                                   'end',
-                                   'end'])
-        expect(cli.run(['--auto-correct'])).to eq(1)
-        expect(IO.read('example.rb'))
-          .to eq(['# encoding: utf-8',
-                  'module Foo',
-                  '  class Bar',
-                  '    # This is a nice long',
-                  '    # comment',
-                  '    # which spans a few lines',
-                  '    def baz',
-                  '      do_something',
-                  '    end',
-                  '  end',
-                  'end',
-                  ''].join("\n"))
-      end
-
-      it 'can correct two problems with blocks' do
-        # {} should be do..end and space is missing.
-        create_file('example.rb', ['# encoding: utf-8',
-                                   '(1..10).each{ |i|',
-                                   '  puts i',
-                                   '}'])
-        expect(cli.run(['--auto-correct'])).to eq(0)
-        expect(IO.read('example.rb'))
-          .to eq(['# encoding: utf-8',
-                  '(1..10).each do |i|',
-                  '  puts i',
-                  'end',
-                  ''].join("\n"))
-      end
-
-      it 'can handle spaces when removing braces' do
-        create_file('example.rb',
-                    ['# encoding: utf-8',
-                     "assert_post_status_code 400, 's', {:type => 'bad'}"])
-        expect(cli.run(%w(--auto-correct --format emacs))).to eq(0)
-        expect(IO.read('example.rb'))
-          .to eq(['# encoding: utf-8',
-                  "assert_post_status_code 400, 's', type: 'bad'",
-                  ''].join("\n"))
-        e = abs('example.rb')
-        expect($stdout.string)
-          .to eq(["#{e}:2:35: C: [Corrected] Redundant curly braces around " \
-                  'a hash parameter.',
-                  "#{e}:2:35: C: [Corrected] Use the new Ruby 1.9 hash " \
-                  'syntax.',
-                  # TODO: Don't report that a problem is corrected when it
-                  # actually went away due to another correction.
-                  "#{e}:2:35: C: [Corrected] Space inside { missing.",
-                  # TODO: Don't report duplicates (HashSyntax in this case).
-                  "#{e}:2:36: C: [Corrected] Use the new Ruby 1.9 hash " \
-                  'syntax.',
-                  "#{e}:2:50: C: [Corrected] Space inside } missing.",
-                  ''].join("\n"))
-      end
-
-      # A case where two cops, EmptyLinesAroundBody and EmptyLines, try to
-      # remove the same line in autocorrect.
-      it 'can correct two empty lines at end of class body' do
-        create_file('example.rb', ['class Test',
-                                   '  def f',
-                                   '  end',
-                                   '',
-                                   '',
-                                   'end'])
-        expect(cli.run(['--auto-correct'])).to eq(1)
-        expect($stderr.string).to eq('')
-        expect(IO.read('example.rb')).to eq(['class Test',
-                                             '  def f',
-                                             '  end',
-                                             'end',
-                                             ''].join("\n"))
-      end
-
-      # A case where WordArray's correction can be clobbered by
-      # AccessModifierIndentation's correction.
-      it 'can correct indentation and another thing' do
-        create_file('example.rb', ['# encoding: utf-8',
-                                   'class Dsl',
-                                   'private',
-                                   '  A = ["git", "path",]',
-                                   'end'])
-        expect(cli.run(%w(--auto-correct --format emacs))).to eq(1)
-        expect(IO.read('example.rb')).to eq(['# encoding: utf-8',
-                                             'class Dsl',
-                                             '  private',
-                                             '',
-                                             '  A = %w(git path)',
-                                             'end',
-                                             ''].join("\n"))
-        e = abs('example.rb')
-        expect($stdout.string)
-          .to eq(["#{e}:2:1: C: Missing top-level class documentation " \
-                  'comment.',
-                  "#{e}:3:1: C: [Corrected] Indent access modifiers like " \
-                  '`private`.',
-                  "#{e}:3:1: C: [Corrected] Keep a blank line before and " \
-                  'after `private`.',
-                  "#{e}:3:3: W: Useless `private` access modifier.",
-                  "#{e}:3:3: C: [Corrected] Keep a blank line before and " \
-                  'after `private`.',
-                  "#{e}:4:7: C: [Corrected] Use `%w` or `%W` " \
-                  'for an array of words.',
-                  "#{e}:4:8: C: [Corrected] Prefer single-quoted strings " \
-                  "when you don't need string interpolation or special " \
-                  'symbols.',
-                  "#{e}:4:15: C: [Corrected] Prefer single-quoted strings " \
-                  "when you don't need string interpolation or special " \
-                  'symbols.',
-                  "#{e}:4:21: C: [Corrected] Avoid comma after the last item " \
-                  'of an array.',
-                  "#{e}:5:7: C: [Corrected] Use `%w` or `%W` " \
-                  'for an array of words.',
-                  "#{e}:5:8: C: [Corrected] Prefer single-quoted strings " \
-                  "when you don't need string interpolation or special " \
-                  'symbols.',
-                  "#{e}:5:15: C: [Corrected] Prefer single-quoted strings " \
-                  "when you don't need string interpolation or special " \
-                  'symbols.',
-                  "#{e}:5:21: C: [Corrected] Avoid comma after the last item " \
-                  'of an array.',
-                  ''].join("\n"))
-      end
-
-      # A case where the same cop could try to correct an offense twice in one
-      # place.
-      it 'can correct empty line inside special form of nested modules' do
-        create_file('example.rb', ['module A module B',
-                                   '',
-                                   'end end'])
-        expect(cli.run(['--auto-correct'])).to eq(1)
-        expect(IO.read('example.rb')).to eq(['module A module B',
-                                             'end end',
-                                             ''].join("\n"))
-        uncorrected = $stdout.string.split($RS).select do |line|
-          line.include?('example.rb:') && !line.include?('[Corrected]')
-        end
-        expect(uncorrected).not_to be_empty # Hence exit code 1.
-      end
-
-      it 'can correct single line methods' do
-        create_file('example.rb', ['# encoding: utf-8',
-                                   'def func1; do_something end # comment',
-                                   'def func2() do_1; do_2; end'])
-        expect(cli.run(%w(--auto-correct --format offenses))).to eq(0)
-        expect(IO.read('example.rb')).to eq(['# encoding: utf-8',
-                                             '# comment',
-                                             'def func1',
-                                             '  do_something',
-                                             'end',
-                                             '',
-                                             'def func2',
-                                             '  do_1',
-                                             '  do_2',
-                                             'end',
-                                             ''].join("\n"))
-        expect($stdout.string).to eq(['',
-                                      '10  Style/TrailingWhitespace',
-                                      '3   Style/Semicolon',
-                                      '3   Style/SingleLineMethods',
-                                      '1   Style/DefWithParentheses',
-                                      '1   Style/EmptyLineBetweenDefs',
-                                      '--',
-                                      '18  Total',
-                                      '',
-                                      ''].join("\n"))
-      end
-
-      # In this example, the auto-correction (changing "raise" to "fail")
-      # creates a new problem (alignment of parameters), which is also
-      # corrected automatically.
-      it 'can correct a problems and the problem it creates' do
-        create_file('example.rb',
-                    ['# encoding: utf-8',
-                     'raise NotImplementedError,',
-                     "      'Method should be overridden in child classes'"])
-        expect(cli.run(['--auto-correct'])).to eq(0)
-        expect(IO.read('example.rb'))
-          .to eq(['# encoding: utf-8',
-                  'fail NotImplementedError,',
-                  "     'Method should be overridden in child classes'",
-                  ''].join("\n"))
-        expect($stdout.string)
-          .to eq(['Inspecting 1 file',
-                  'C',
-                  '',
-                  'Offenses:',
-                  '',
-                  'example.rb:2:1: C: [Corrected] Use fail instead of ' \
-                  'raise to signal exceptions.',
-                  'raise NotImplementedError,',
-                  '^^^^^',
-                  'example.rb:3:7: C: [Corrected] Align the parameters of a ' \
-                  'method call if they span more than one line.',
-                  "      'Method should be overridden in child classes'",
-                  '      ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^',
-                  '',
-                  '1 file inspected, 2 offenses detected, 2 offenses ' \
-                  'corrected',
-                  ''].join("\n"))
-      end
-
-      # Thanks to repeated auto-correction, we can get rid of the trailing
-      # spaces, and then the extra empty line.
-      it 'can correct two problems in the same place' do
-        create_file('example.rb',
-                    ['# encoding: utf-8',
-                     '# Example class.',
-                     'class Klass',
-                     '  ',
-                     '  def f',
-                     '  end',
-                     'end'])
-        expect(cli.run(['--auto-correct'])).to eq(0)
-        expect(IO.read('example.rb'))
-          .to eq(['# encoding: utf-8',
-                  '# Example class.',
-                  'class Klass',
-                  '  def f',
-                  '  end',
-                  'end',
-                  ''].join("\n"))
-        expect($stderr.string).to eq('')
-        expect($stdout.string)
-          .to eq(['Inspecting 1 file',
-                  'C',
-                  '',
-                  'Offenses:',
-                  '',
-                  'example.rb:4:1: C: [Corrected] Extra empty line detected ' \
-                  'at class body beginning.',
-                  'example.rb:4:1: C: [Corrected] Trailing whitespace ' \
-                  'detected.',
-                  '',
-                  '1 file inspected, 2 offenses detected, 2 offenses ' \
-                  'corrected',
-                  ''].join("\n"))
-      end
-
-      it 'can correct MethodDefParentheses and other offense' do
-        create_file('example.rb',
-                    ['# encoding: utf-8',
-                     'def primes limit',
-                     '  1.upto(limit).select { |i| i.even? }',
-                     'end'])
-        expect(cli.run(%w(-D --auto-correct))).to eq(0)
-        expect($stderr.string).to eq('')
-        expect(IO.read('example.rb'))
-          .to eq(['# encoding: utf-8',
-                  'def primes(limit)',
-                  '  1.upto(limit).select(&:even?)',
-                  'end',
-                  ''].join("\n"))
-        expect($stdout.string)
-          .to eq(['Inspecting 1 file',
-                  'C',
-                  '',
-                  'Offenses:',
-                  '',
-                  'example.rb:2:12: C: [Corrected] ' \
-                  'Style/MethodDefParentheses: ' \
-                  'Use def with parentheses when there are parameters.',
-                  'def primes limit',
-                  '           ^^^^^',
-                  'example.rb:3:24: C: [Corrected] Style/SymbolProc: ' \
-                  'Pass &:even? as an argument to select instead of a block.',
-                  '  1.upto(limit).select { |i| i.even? }',
-                  '                       ^^^^^^^^^^^^^^^',
-                  '',
-                  '1 file inspected, 2 offenses detected, 2 offenses ' \
-                  'corrected',
-                  ''].join("\n"))
-      end
-
-      it 'can correct WordArray and SpaceAfterComma offenses' do
-        create_file('example.rb',
-                    ['# encoding: utf-8',
-                     "f(type: ['offline','offline_payment'],",
-                     "  bar_colors: ['958c12','953579','ff5800','0085cc'])"])
-        expect(cli.run(%w(-D --auto-correct --format o))).to eq(0)
-        expect($stdout.string)
-          .to eq(['',
-                  '4  Style/SpaceAfterComma',
-                  '2  Style/WordArray',
-                  '--',
-                  '6  Total',
-                  '',
-                  ''].join("\n"))
-        expect(IO.read('example.rb'))
-          .to eq(['# encoding: utf-8',
-                  'f(type: %w(offline offline_payment),',
-                  '  bar_colors: %w(958c12 953579 ff5800 0085cc))',
-                  ''].join("\n"))
-      end
-
-      it 'can correct SpaceAfterComma and HashSyntax offenses' do
-        create_file('example.rb',
-                    ['# encoding: utf-8',
-                     "I18n.t('description',:property_name => property.name)"])
-        expect(cli.run(%w(-D --auto-correct --format emacs))).to eq(0)
-        expect($stdout.string)
-          .to eq(["#{abs('example.rb')}:2:21: C: [Corrected] " \
-                  'Style/SpaceAfterComma: Space missing after comma.',
-                  "#{abs('example.rb')}:2:22: C: [Corrected] " \
-                  'Style/HashSyntax: Use the new Ruby 1.9 hash syntax.',
-                  ''].join("\n"))
-        expect(IO.read('example.rb'))
-          .to eq(['# encoding: utf-8',
-                  "I18n.t('description', property_name: property.name)",
-                  ''].join("\n"))
-      end
-
-      it 'can correct HashSyntax and SpaceAroundOperators offenses' do
-        create_file('example.rb',
-                    ['# encoding: utf-8',
-                     '{ :b=>1 }'])
-        expect(cli.run(%w(-D --auto-correct --format emacs))).to eq(0)
-        expect(IO.read('example.rb')).to eq(['# encoding: utf-8',
-                                             '{ b: 1 }',
-                                             ''].join("\n"))
-        expect($stdout.string)
-          .to eq(["#{abs('example.rb')}:2:3: C: [Corrected] " \
-                  'Style/HashSyntax: Use the new Ruby 1.9 hash syntax.',
-                  "#{abs('example.rb')}:2:5: C: [Corrected] " \
-                  'Style/SpaceAroundOperators: Surrounding space missing for ' \
-                  'operator `=>`.',
-                  ''].join("\n"))
-      end
-
-      it 'can correct HashSyntax when --only is used' do
-        create_file('example.rb',
-                    ['# encoding: utf-8',
-                     '{ :b=>1 }'])
-        expect(cli.run(%w(--auto-correct -f emacs
-                          --only Style/HashSyntax))).to eq(0)
-        expect($stderr.string).to eq('')
-        expect(IO.read('example.rb')).to eq(['# encoding: utf-8',
-                                             '{ b: 1 }',
-                                             ''].join("\n"))
-        expect($stdout.string)
-          .to eq(["#{abs('example.rb')}:2:3: C: [Corrected] Use the new " \
-                  'Ruby 1.9 hash syntax.',
-                  ''].join("\n"))
-      end
-
-      it 'can correct TrailingBlankLines and TrailingWhitespace offenses' do
-        create_file('example.rb',
-                    ['# encoding: utf-8',
-                     '',
-                     '  ',
-                     '',
-                     ''])
-        expect(cli.run(%w(--auto-correct --format emacs))).to eq(0)
-        expect(IO.read('example.rb')).to eq(['# encoding: utf-8',
-                                             ''].join("\n"))
-        expect($stdout.string)
-          .to eq(["#{abs('example.rb')}:2:1: C: [Corrected] 3 trailing " \
-                  'blank lines detected.',
-                  "#{abs('example.rb')}:3:1: C: [Corrected] Trailing " \
-                  'whitespace detected.',
-                  ''].join("\n"))
-      end
-
-      it 'can correct MethodCallParentheses and EmptyLiteral offenses' do
-        create_file('example.rb',
-                    ['# encoding: utf-8',
-                     'Hash.new()'])
-        expect(cli.run(%w(--auto-correct --format emacs))).to eq(0)
-        expect($stderr.string).to eq('')
-        expect(IO.read('example.rb')).to eq(['# encoding: utf-8',
-                                             '{}',
-                                             ''].join("\n"))
-        expect($stdout.string)
-          .to eq(["#{abs('example.rb')}:2:1: C: [Corrected] Use hash " \
-                  'literal `{}` instead of `Hash.new`.',
-                  "#{abs('example.rb')}:2:9: C: [Corrected] Do not use " \
-                  'parentheses for method calls with no arguments.',
-                  ''].join("\n"))
-      end
-
-      it 'can correct IndentHash offenses with separator style' do
-        create_file('example.rb',
-                    ['# encoding: utf-8',
-                     'CONVERSION_CORRESPONDENCE = {',
-                     '              match_for_should: :match,',
-                     '          match_for_should_not: :match_when_negated,',
-                     '    failure_message_for_should: :failure_message,',
-                     'failure_message_for_should_not: :failure_message_when',
-                     '}'])
-        create_file('.rubocop.yml',
-                    ['Style/AlignHash:',
-                     '  EnforcedColonStyle: separator'])
-        expect(cli.run(%w(--auto-correct))).to eq(0)
-        expect(IO.read('example.rb'))
-          .to eq(['# encoding: utf-8',
-                  'CONVERSION_CORRESPONDENCE = {',
-                  '                match_for_should: :match,',
-                  '            match_for_should_not: :match_when_negated,',
-                  '      failure_message_for_should: :failure_message,',
-                  '  failure_message_for_should_not: :failure_message_when',
-                  '}',
-                  ''].join("\n"))
-      end
-
-      it 'does not say [Corrected] if correction was avoided' do
-        src = ['# encoding: utf-8',
-               'not a && b',
-               'func a do b end',
-               "Signal.trap('TERM') { system(cmd); exit }"]
-        corrected = ['# encoding: utf-8',
-                     'not a && b',
-                     'func a do b end',
-                     "Signal.trap('TERM') { system(cmd); exit }"]
-        offenses =
-          ['== example.rb ==',
-           'C:  2:  1: Use ! instead of not.',
-           'C:  3:  8: Prefer {...} over do...end for single-line blocks.',
-           'C:  4: 34: Do not use semicolons to terminate expressions.']
-
-        if RUBY_VERSION >= '2'
-          src += ['def self.some_method(foo, bar: 1)',
-                  '  log.debug(foo)',
-                  'end']
-          corrected += ['def self.some_method(foo, bar: 1)',
-                        '  log.debug(foo)',
-                        'end']
-          offenses += ['W:  5: 27: Unused method argument - bar.']
-          summary = '1 file inspected, 4 offenses detected'
-        else
-          summary = '1 file inspected, 3 offenses detected'
-        end
-        create_file('example.rb', src)
-        expect(cli.run(%w(-a -f simple))).to eq(1)
-        expect($stderr.string).to eq('')
-        expect(IO.read('example.rb')).to eq(corrected.join("\n") + "\n")
-        expect($stdout.string)
-          .to eq((offenses + ['', summary, '']).join("\n"))
-      end
-
-      it 'does not hang SpaceAfterPunctuation and SpaceInsideParens' do
-        create_file('example.rb',
-                    ['# encoding: utf-8',
-                     'some_method(a, )'])
-        Timeout.timeout(10) do
-          expect(cli.run(%w(--auto-correct))).to eq(0)
-        end
-        expect($stderr.string).to eq('')
-        expect(IO.read('example.rb')).to eq(['# encoding: utf-8',
-                                             'some_method(a)',
-                                             ''].join("\n"))
-      end
-
-      it 'does not hang SpaceAfterPunctuation and SpaceInsideBrackets' do
-        create_file('example.rb',
-                    ['# encoding: utf-8',
-                     'puts [1, ]'])
-        Timeout.timeout(10) do
-          expect(cli.run(%w(--auto-correct))).to eq(0)
-        end
-        expect($stderr.string).to eq('')
-        expect(IO.read('example.rb')).to eq(['# encoding: utf-8',
-                                             'puts [1]',
-                                             ''].join("\n"))
-      end
-
-      it 'can be disabled for any cop in configuration' do
-        create_file('example.rb', ['# encoding: utf-8',
-                                   'puts "Hello", 123456'])
-        create_file('.rubocop.yml', ['Style/StringLiterals:',
-                                     '  AutoCorrect: false'])
-        expect(cli.run(%w(--auto-correct))).to eq(1)
-        expect($stderr.string).to eq('')
-        expect(IO.read('example.rb')).to eq(['# encoding: utf-8',
-                                             'puts "Hello", 123_456',
-                                             ''].join("\n"))
-      end
-    end
-
-    describe '--auto-gen-config' do
-      before(:each) do
-        RuboCop::Formatter::DisabledConfigFormatter
-          .config_to_allow_offenses = {}
-      end
-
-      it 'overwrites an existing todo file' do
-        create_file('example1.rb', ['# encoding: utf-8',
-                                    'x= 0 ',
-                                    '#' * 85,
-                                    'y ',
-                                    'puts x'])
-        create_file('.rubocop_todo.yml', ['Metrics/LineLength:',
-                                          '  Enabled: false'])
-        create_file('.rubocop.yml', ['inherit_from: .rubocop_todo.yml'])
-        expect(cli.run(['--auto-gen-config'])).to eq(1)
-        expect(IO.readlines('.rubocop_todo.yml')[8..-1].map(&:chomp))
-          .to eq(['# Offense count: 1',
-                  '# Configuration parameters: AllowHeredoc, AllowURI, ' \
-                  'URISchemes.',
-                  '# URISchemes: http, https',
-                  'Metrics/LineLength:',
-                  '  Max: 85',
-                  '',
-                  '# Offense count: 1',
-                  '# Cop supports --auto-correct.',
-                  '# Configuration parameters: AllowForAlignment.',
-                  'Style/SpaceAroundOperators:',
-                  '  Exclude:',
-                  "    - 'example1.rb'",
-                  '',
-                  '# Offense count: 2',
-                  '# Cop supports --auto-correct.',
-                  'Style/TrailingWhitespace:',
-                  '  Exclude:',
-                  "    - 'example1.rb'"])
-
-        # Create new CLI instance to avoid using cached configuration.
-        new_cli = described_class.new
-
-        expect(new_cli.run(['example1.rb'])).to eq(0)
-      end
-
-      it 'honors rubocop:disable comments' do
-        create_file('example1.rb', ['# encoding: utf-8',
-                                    '#' * 81,
-                                    '# rubocop:disable LineLength',
-                                    '#' * 85,
-                                    'y ',
-                                    'puts 123456'])
-        create_file('.rubocop.yml', ['inherit_from: .rubocop_todo.yml'])
-        expect(cli.run(['--auto-gen-config'])).to eq(1)
-        expect(IO.readlines('.rubocop_todo.yml')[8..-1].join)
-          .to eq(['# Offense count: 1',
-                  '# Configuration parameters: AllowHeredoc, AllowURI, ' \
-                  'URISchemes.',
-                  '# URISchemes: http, https',
-                  'Metrics/LineLength:',
-                  '  Max: 81',
-                  '',
-                  '# Offense count: 1',
-                  '# Cop supports --auto-correct.',
-                  'Style/NumericLiterals:',
-                  '  MinDigits: 7',
-                  '',
-                  '# Offense count: 1',
-                  '# Cop supports --auto-correct.',
-                  'Style/TrailingWhitespace:',
-                  '  Exclude:',
-                  "    - 'example1.rb'",
-                  ''].join("\n"))
-      end
-
-      it 'can generate a todo list' do
-        create_file('example1.rb', ['# encoding: utf-8',
-                                    '$x= 0 ',
-                                    '#' * 90,
-                                    '#' * 85,
-                                    'y ',
-                                    'puts x'])
-        create_file('example2.rb', ['# encoding: utf-8',
-                                    "\tx = 0",
-                                    'puts x',
-                                    '',
-                                    'class A',
-                                    '  def a',
-                                    '  end',
-                                    'end'])
-        # Make ConfigLoader reload the default configuration so that its
-        # absolute Exclude paths will point into this example's work directory.
-        RuboCop::ConfigLoader.default_configuration = nil
-
-        expect(cli.run(['--auto-gen-config'])).to eq(1)
-        expect($stderr.string).to eq('')
-        expect($stdout.string)
-          .to include(['Created .rubocop_todo.yml.',
-                       'Run `rubocop --config .rubocop_todo.yml`, or ' \
-                       'add `inherit_from: .rubocop_todo.yml` in a ' \
-                       '.rubocop.yml file.',
-                       ''].join("\n"))
-        expected =
-          ['# This configuration was generated by',
-           '# `rubocop --auto-gen-config`',
-           /# on .* using RuboCop version .*/,
-           '# The point is for the user to remove these configuration records',
-           '# one by one as the offenses are removed from the code base.',
-           '# Note that changes in the inspected code, or installation of new',
-           '# versions of RuboCop, may require this file to be generated ' \
-           'again.',
-           '',
-           '# Offense count: 2',
-           '# Configuration parameters: AllowHeredoc, AllowURI, URISchemes.',
-           '# URISchemes: http, https',
-           'Metrics/LineLength:',
-           '  Max: 90',
-           '',
-           '# Offense count: 1',
-           '# Cop supports --auto-correct.',
-           'Style/CommentIndentation:',
-           '  Exclude:',
-           "    - 'example2.rb'",
-           '',
-           '# Offense count: 1',
-           'Style/Documentation:',
-           '  Exclude:',
-           "    - 'spec/**/*'", # Copied from default configuration
-           "    - 'test/**/*'", # Copied from default configuration
-           "    - 'example2.rb'",
-           '',
-           '# Offense count: 1',
-           '# Configuration parameters: AllowedVariables.',
-           'Style/GlobalVars:',
-           '  Exclude:',
-           "    - 'example1.rb'",
-           '',
-           '# Offense count: 2',
-           '# Cop supports --auto-correct.',
-           '# Configuration parameters: EnforcedStyle, SupportedStyles.',
-           '# SupportedStyles: normal, rails',
-           'Style/IndentationConsistency:',
-           '  Exclude:',
-           "    - 'example2.rb'",
-           '',
-           '# Offense count: 1',
-           '# Cop supports --auto-correct.',
-           'Style/InitialIndentation:',
-           '  Exclude:',
-           "    - 'example2.rb'",
-           '',
-           '# Offense count: 1',
-           '# Cop supports --auto-correct.',
-           '# Configuration parameters: AllowForAlignment.',
-           'Style/SpaceAroundOperators:',
-           '  Exclude:',
-           "    - 'example1.rb'",
-           '',
-           '# Offense count: 1',
-           '# Cop supports --auto-correct.',
-           'Style/Tab:',
-           '  Exclude:',
-           "    - 'example2.rb'",
-           '',
-           '# Offense count: 2',
-           '# Cop supports --auto-correct.',
-           'Style/TrailingWhitespace:',
-           '  Exclude:',
-           "    - 'example1.rb'"]
-        actual = IO.read('.rubocop_todo.yml').split($RS)
-        expected.each_with_index do |line, ix|
-          if line.is_a?(String)
-            expect(actual[ix]).to eq(line)
-          else
-            expect(actual[ix]).to match(line)
-          end
-        end
-      end
-
-      it 'can generate Exclude properties with a given limit' do
-        create_file('example1.rb', ['# encoding: utf-8',
-                                    '$x= 0 ',
-                                    '#' * 90,
-                                    '#' * 85,
-                                    'y ',
-                                    'puts x'])
-        create_file('example2.rb', ['# encoding: utf-8',
-                                    '#' * 85,
-                                    "\tx = 0",
-                                    'puts x'])
-        expect(cli.run(['--auto-gen-config', '--exclude-limit', '1'])).to eq(1)
-        expected =
-          ['# This configuration was generated by',
-           '# `rubocop --auto-gen-config --exclude-limit 1`',
-           /# on .* using RuboCop version .*/,
-           '# The point is for the user to remove these configuration records',
-           '# one by one as the offenses are removed from the code base.',
-           '# Note that changes in the inspected code, or installation of new',
-           '# versions of RuboCop, may require this file to be generated ' \
-           'again.',
-           '',
-           '# Offense count: 3',
-           '# Configuration parameters: AllowHeredoc, AllowURI, URISchemes.',
-           '# URISchemes: http, https',
-           'Metrics/LineLength:',
-           '  Max: 90', # Offense occurs in 2 files, limit is 1, so no Exclude.
-           '',
-           '# Offense count: 1',
-           '# Cop supports --auto-correct.',
-           'Style/CommentIndentation:',
-           '  Exclude:',
-           "    - 'example2.rb'",
-           '',
-           '# Offense count: 1',
-           '# Configuration parameters: AllowedVariables.',
-           'Style/GlobalVars:',
-           '  Exclude:',
-           "    - 'example1.rb'",
-           '',
-           '# Offense count: 1',
-           '# Cop supports --auto-correct.',
-           '# Configuration parameters: EnforcedStyle, SupportedStyles.',
-           '# SupportedStyles: normal, rails',
-           'Style/IndentationConsistency:',
-           '  Exclude:',
-           "    - 'example2.rb'",
-           '',
-           '# Offense count: 1',
-           '# Cop supports --auto-correct.',
-           'Style/InitialIndentation:',
-           '  Exclude:',
-           "    - 'example2.rb'",
-           '',
-           '# Offense count: 1',
-           '# Cop supports --auto-correct.',
-           '# Configuration parameters: AllowForAlignment.',
-           'Style/SpaceAroundOperators:',
-           '  Exclude:',
-           "    - 'example1.rb'",
-           '',
-           '# Offense count: 1',
-           '# Cop supports --auto-correct.',
-           'Style/Tab:',
-           '  Exclude:',
-           "    - 'example2.rb'",
-           '',
-           '# Offense count: 2',
-           '# Cop supports --auto-correct.',
-           'Style/TrailingWhitespace:',
-           '  Exclude:',
-           "    - 'example1.rb'"]
-        actual = IO.read('.rubocop_todo.yml').split($RS)
-        expected.each_with_index do |line, ix|
-          if line.is_a?(String)
-            expect(actual[ix]).to eq(line)
-          else
-            expect(actual[ix]).to match(line)
-          end
-        end
-      end
-
-      it 'does not generate configuration for the Syntax cop' do
-        create_file('example1.rb', ['# encoding: utf-8',
-                                    'x = < ', # Syntax error
-                                    'puts x'])
-        create_file('example2.rb', ['# encoding: utf-8',
-                                    "\tx = 0",
-                                    'puts x'])
-        expect(cli.run(['--auto-gen-config'])).to eq(1)
-        expect($stderr.string).to eq('')
-        expected =
-          ['# This configuration was generated by',
-           '# `rubocop --auto-gen-config`',
-           /# on .* using RuboCop version .*/,
-           '# The point is for the user to remove these configuration records',
-           '# one by one as the offenses are removed from the code base.',
-           '# Note that changes in the inspected code, or installation of new',
-           '# versions of RuboCop, may require this file to be generated ' \
-           'again.',
-           '',
-           '# Offense count: 1',
-           '# Cop supports --auto-correct.',
-           'Style/CommentIndentation:',
-           '  Exclude:',
-           "    - 'example2.rb'",
-           '',
-           '# Offense count: 1',
-           '# Cop supports --auto-correct.',
-           '# Configuration parameters: EnforcedStyle, SupportedStyles.',
-           '# SupportedStyles: normal, rails',
-           'Style/IndentationConsistency:',
-           '  Exclude:',
-           "    - 'example2.rb'",
-           '',
-           '# Offense count: 1',
-           '# Cop supports --auto-correct.',
-           'Style/InitialIndentation:',
-           '  Exclude:',
-           "    - 'example2.rb'",
-           '',
-           '# Offense count: 1',
-           '# Cop supports --auto-correct.',
-           'Style/Tab:',
-           '  Exclude:',
-           "    - 'example2.rb'"]
-        actual = IO.read('.rubocop_todo.yml').split($RS)
-        expect(actual.length).to eq(expected.length)
-        expected.each_with_index do |line, ix|
-          if line.is_a?(String)
-            expect(actual[ix]).to eq(line)
-          else
-            expect(actual[ix]).to match(line)
-          end
-        end
-      end
-
-      it 'generates a todo list that removes the reports' do
-        create_file('example.rb', ['# encoding: utf-8',
-                                   'y.gsub!(/abc\/xyz/, x)'])
-        expect(cli.run(%w(--format emacs))).to eq(1)
-        expect($stdout.string).to eq("#{abs('example.rb')}:2:9: C: Use `%r` " \
-                                     "around regular expression.\n")
-        expect(cli.run(['--auto-gen-config'])).to eq(1)
-        expected =
-          ['# This configuration was generated by',
-           '# `rubocop --auto-gen-config`',
-           /# on .* using RuboCop version .*/,
-           '# The point is for the user to remove these configuration records',
-           '# one by one as the offenses are removed from the code base.',
-           '# Note that changes in the inspected code, or installation of new',
-           '# versions of RuboCop, may require this file to be generated ' \
-           'again.',
-           '',
-           '# Offense count: 1',
-           '# Cop supports --auto-correct.',
-           '# Configuration parameters: EnforcedStyle, SupportedStyles, ' \
-           'AllowInnerSlashes.',
-           '# SupportedStyles: slashes, percent_r, mixed',
-           'Style/RegexpLiteral:',
-           '  Exclude:',
-           "    - 'example.rb'"]
-        actual = IO.read('.rubocop_todo.yml').split($RS)
-        expected.each_with_index do |line, ix|
-          if line.is_a?(String)
-            expect(actual[ix]).to eq(line)
-          else
-            expect(actual[ix]).to match(line)
-          end
-        end
-        $stdout = StringIO.new
-        result = cli.run(%w(--config .rubocop_todo.yml --format emacs))
-        expect($stdout.string).to eq('')
-        expect(result).to eq(0)
-      end
-
-      it 'does not include offense counts when --no-offense-counts is used' do
-        create_file('example1.rb', ['# encoding: utf-8',
-                                    '$x= 0 ',
-                                    '#' * 90,
-                                    '#' * 85,
-                                    'y ',
-                                    'puts x'])
-        create_file('example2.rb', ['# encoding: utf-8',
-                                    "\tx = 0",
-                                    'puts x',
-                                    '',
-                                    'class A',
-                                    '  def a',
-                                    '  end',
-                                    'end'])
-        # Make ConfigLoader reload the default configuration so that its
-        # absolute Exclude paths will point into this example's work directory.
-        RuboCop::ConfigLoader.default_configuration = nil
-
-        expect(cli.run(['--auto-gen-config', '--no-offense-counts'])).to eq(1)
-        expect($stderr.string).to eq('')
-        expect($stdout.string)
-          .to include(['Created .rubocop_todo.yml.',
-                       'Run `rubocop --config .rubocop_todo.yml`, or ' \
-                       'add `inherit_from: .rubocop_todo.yml` in a ' \
-                       '.rubocop.yml file.',
-                       ''].join("\n"))
-        expected =
-          ['# This configuration was generated by',
-           '# `rubocop --auto-gen-config`',
-           /# on .* using RuboCop version .*/,
-           '# The point is for the user to remove these configuration records',
-           '# one by one as the offenses are removed from the code base.',
-           '# Note that changes in the inspected code, or installation of new',
-           '# versions of RuboCop, may require this file to be generated ' \
-           'again.',
-           '',
-           '# Configuration parameters: AllowHeredoc, AllowURI, URISchemes.',
-           '# URISchemes: http, https',
-           'Metrics/LineLength:',
-           '  Max: 90',
-           '',
-           '# Cop supports --auto-correct.',
-           'Style/CommentIndentation:',
-           '  Exclude:',
-           "    - 'example2.rb'",
-           '',
-           'Style/Documentation:',
-           '  Exclude:',
-           "    - 'spec/**/*'", # Copied from default configuration
-           "    - 'test/**/*'", # Copied from default configuration
-           "    - 'example2.rb'",
-           '',
-           '# Configuration parameters: AllowedVariables.',
-           'Style/GlobalVars:',
-           '  Exclude:',
-           "    - 'example1.rb'",
-           '',
-           '# Cop supports --auto-correct.',
-           '# Configuration parameters: EnforcedStyle, SupportedStyles.',
-           '# SupportedStyles: normal, rails',
-           'Style/IndentationConsistency:',
-           '  Exclude:',
-           "    - 'example2.rb'",
-           '',
-           '# Cop supports --auto-correct.',
-           'Style/InitialIndentation:',
-           '  Exclude:',
-           "    - 'example2.rb'",
-           '',
-           '# Cop supports --auto-correct.',
-           '# Configuration parameters: AllowForAlignment.',
-           'Style/SpaceAroundOperators:',
-           '  Exclude:',
-           "    - 'example1.rb'",
-           '',
-           '# Cop supports --auto-correct.',
-           'Style/Tab:',
-           '  Exclude:',
-           "    - 'example2.rb'",
-           '',
-           '# Cop supports --auto-correct.',
-           'Style/TrailingWhitespace:',
-           '  Exclude:',
-           "    - 'example1.rb'"]
-        actual = IO.read('.rubocop_todo.yml').split($RS)
-        expected.each_with_index do |line, ix|
-          if line.is_a?(String)
-            expect(actual[ix]).to eq(line)
-          else
-            expect(actual[ix]).to match(line)
-          end
-        end
-      end
-    end
-
-    describe '--only' do
-      context 'when one cop is given' do
-        it 'runs just one cop' do
-          # The disable comment should not be reported as unnecessary (even if
-          # it is) since --only overrides configuration.
-          create_file('example.rb', ['# rubocop:disable LineLength',
-                                     'if x== 0 ',
-                                     "\ty",
-                                     'end'])
-          # IfUnlessModifier depends on the configuration of LineLength.
-
-          expect(cli.run(['--format', 'simple',
-                          '--only', 'Style/IfUnlessModifier',
-                          'example.rb'])).to eq(1)
-          expect($stdout.string)
-            .to eq(['== example.rb ==',
-                    'C:  2:  1: Favor modifier if usage when ' \
-                    'having a single-line body. Another good alternative is ' \
-                    'the usage of control flow &&/||.',
-                    '',
-                    '1 file inspected, 1 offense detected',
-                    ''].join("\n"))
-        end
-
-        it 'exits with error if an incorrect cop name is passed' do
-          create_file('example.rb', ['if x== 0 ',
-                                     "\ty",
-                                     'end'])
-          expect(cli.run(['--only', 'Style/123'])).to eq(1)
-          expect($stderr.string)
-            .to include('Unrecognized cop or namespace: Style/123.')
-        end
-
-        it 'exits with error if an empty string is given' do
-          create_file('example.rb', 'x')
-          expect(cli.run(['--only', ''])).to eq(1)
-          expect($stderr.string).to include('Unrecognized cop or namespace: .')
-        end
-
-        %w(Syntax Lint/Syntax).each do |name|
-          it "only checks syntax if #{name} is given" do
-            create_file('example.rb', 'x ')
-            expect(cli.run(['--only', name])).to eq(0)
-            expect($stdout.string).to include('no offenses detected')
-          end
-        end
-
-        %w(Lint/UnneededDisable UnneededDisable).each do |name|
-          it "exits with error if cop name #{name} is passed" do
-            create_file('example.rb', ['if x== 0 ',
-                                       "\ty",
-                                       'end'])
-            expect(cli.run(['--only', 'UnneededDisable'])).to eq(1)
-            expect($stderr.string)
-              .to include('Lint/UnneededDisable can not be used with --only.')
-          end
-        end
-
-        it 'accepts cop names from plugins' do
-          create_file('.rubocop.yml', ['require: rubocop_ext',
-                                       '',
-                                       'Style/SomeCop:',
-                                       '  Description: Something',
-                                       '  Enabled: true'])
-          create_file('rubocop_ext.rb', ['module RuboCop',
-                                         '  module Cop',
-                                         '    module Style',
-                                         '      class SomeCop < Cop',
-                                         '      end',
-                                         '    end',
-                                         '  end',
-                                         'end'])
-          create_file('redirect.rb', '$stderr = STDOUT')
-          rubocop = "#{RuboCop::ConfigLoader::RUBOCOP_HOME}/bin/rubocop"
-          # Since we define a new cop class, we have to do this in a separate
-          # process. Otherwise, the extra cop will affect other specs.
-          output =
-            `ruby -I . #{rubocop} --require redirect.rb --only Style/SomeCop`
-          expect($CHILD_STATUS.success?).to be_truthy
-          # The warning about the unrecognized cop is expected. It's given due
-          # to the fact that we haven't supplied any default configuration for
-          # rubocop_ext in this example.
-          expect(output)
-            .to eq(['Inspecting 2 files',
-                    '..',
-                    '',
-                    '2 files inspected, no offenses detected',
-                    ''].join("\n"))
-        end
-
-        context 'without using namespace' do
-          it 'runs just one cop' do
-            create_file('example.rb', ['if x== 0 ',
-                                       "\ty",
-                                       'end'])
-
-            expect(cli.run(['--format', 'simple',
-                            '--display-cop-names',
-                            '--only', 'IfUnlessModifier',
-                            'example.rb'])).to eq(1)
-            expect($stdout.string)
-              .to eq(['== example.rb ==',
-                      'C:  1:  1: Style/IfUnlessModifier: Favor modifier if ' \
-                      'usage when having a single-line body. Another good ' \
-                      'alternative is the usage of control flow &&/||.',
-                      '',
-                      '1 file inspected, 1 offense detected',
-                      ''].join("\n"))
-          end
-        end
-
-        it 'enables the given cop' do
-          create_file('example.rb',
-                      ['x = 0 ',
-                       # Disabling comments still apply.
-                       '# rubocop:disable Style/TrailingWhitespace',
-                       'y = 1  '])
-
-          create_file('.rubocop.yml', ['Style/TrailingWhitespace:',
-                                       '  Enabled: false'])
-
-          expect(cli.run(['--format', 'simple',
-                          '--only', 'Style/TrailingWhitespace',
-                          'example.rb'])).to eq(1)
-          expect($stderr.string).to eq('')
-          expect($stdout.string)
-            .to eq(['== example.rb ==',
-                    'C:  1:  6: Trailing whitespace detected.',
-                    '',
-                    '1 file inspected, 1 offense detected',
-                    ''].join("\n"))
-        end
-      end
-
-      context 'when several cops are given' do
-        it 'runs the given cops' do
-          create_file('example.rb', ['if x== 100000000000000 ',
-                                     "\ty",
-                                     'end'])
-          expect(cli.run(['--format', 'simple',
-                          '--only',
-                          'Style/IfUnlessModifier,Style/Tab,' \
-                          'Style/SpaceAroundOperators',
-                          'example.rb'])).to eq(1)
-          expect($stderr.string).to eq('')
-          expect($stdout.string)
-            .to eq(['== example.rb ==',
-                    'C:  1:  1: Favor modifier if usage when ' \
-                    'having a single-line body. Another good alternative is ' \
-                    'the usage of control flow &&/||.',
-                    'C:  1:  5: Surrounding space missing for operator ==.',
-                    'C:  2:  1: Tab detected.',
-                    '',
-                    '1 file inspected, 3 offenses detected',
-                    ''].join("\n"))
-        end
-
-        context 'and --lint' do
-          it 'runs the given cops plus all enabled lint cops' do
-            create_file('example.rb', ['if x== 100000000000000 ',
-                                       "\ty = 3",
-                                       '  end'])
-            create_file('.rubocop.yml', ['Lint/EndAlignment:',
-                                         '  Enabled: false'])
-            expect(cli.run(['--format', 'simple',
-                            '--only', 'Style/Tab,Style/SpaceAroundOperators',
-                            '--lint',
-                            'example.rb'])).to eq(1)
-            expect($stdout.string)
-              .to eq(['== example.rb ==',
-                      'C:  1:  5: Surrounding space missing for operator ==.',
-                      'C:  2:  1: Tab detected.',
-                      'W:  2:  2: Useless assignment to variable - y.',
-                      '',
-                      '1 file inspected, 3 offenses detected',
-                      ''].join("\n"))
-          end
-        end
-      end
-
-      context 'when a namespace is given' do
-        it 'runs all enabled cops in that namespace' do
-          create_file('example.rb', ['if x== 100000000000000 ',
-                                     '  ' + '#' * 100,
-                                     "\ty",
-                                     'end'])
-          expect(cli.run(%w(-f offenses --only Metrics example.rb))).to eq(1)
-          expect($stdout.string).to eq(['',
-                                        '1  Metrics/LineLength',
-                                        '--',
-                                        '1  Total',
-                                        '',
-                                        ''].join("\n"))
-        end
-      end
-
-      context 'when two namespaces are given' do
-        it 'runs all enabled cops in those namespaces' do
-          create_file('example.rb', ['# encoding: utf-8',
-                                     'if x== 100000000000000 ',
-                                     '  # ' + '-' * 98,
-                                     "\ty",
-                                     'end'])
-          create_file('.rubocop.yml', ['Style/SpaceAroundOperators:',
-                                       '  Enabled: false'])
-          expect(cli.run(%w(-f o --only Metrics,Style example.rb))).to eq(1)
-          expect($stdout.string)
-            .to eq(['',
-                    '1  Metrics/LineLength',
-                    '1  Style/CommentIndentation',
-                    '1  Style/IndentationWidth',
-                    '1  Style/NumericLiterals',
-                    '1  Style/Tab',
-                    '1  Style/TrailingWhitespace',
-                    '--',
-                    '6  Total',
-                    '',
-                    ''].join("\n"))
-        end
-      end
-    end
-
-    describe '--except' do
-      context 'when one name is given' do
-        it 'exits with error if the cop name is incorrect' do
-          create_file('example.rb', ['if x== 0 ',
-                                     "\ty",
-                                     'end'])
-          expect(cli.run(['--except', 'Style/123'])).to eq(1)
-          expect($stderr.string)
-            .to include('Unrecognized cop or namespace: Style/123.')
-        end
-
-        it 'exits with error if an empty string is given' do
-          create_file('example.rb', 'x')
-          expect(cli.run(['--except', ''])).to eq(1)
-          expect($stderr.string).to include('Unrecognized cop or namespace: .')
-        end
-
-        %w(Syntax Lint/Syntax).each do |name|
-          it "exits with error if #{name} is given" do
-            create_file('example.rb', 'x ')
-            expect(cli.run(['--except', name])).to eq(1)
-            expect($stderr.string)
-              .to include('Syntax checking can not be turned off.')
-          end
-        end
-      end
-
-      context 'when one cop plus one namespace are given' do
-        it 'runs all cops except the given' do
-          # The disable comment should not be reported as unnecessary (even if
-          # it is) since --except overrides configuration.
-          create_file('example.rb', ['# rubocop:disable LineLength',
-                                     'if x== 0 ',
-                                     "\ty = 3",
-                                     'end'])
-          expect(cli.run(['--format', 'offenses',
-                          '--except', 'Style/IfUnlessModifier,Lint',
-                          'example.rb'])).to eq(1)
-          expect($stdout.string)
-            .to eq(['',
-                    # Note: No Lint/UselessAssignment offense.
-                    '1  Style/IndentationWidth',
-                    '1  Style/SpaceAroundOperators',
-                    '1  Style/Tab',
-                    '1  Style/TrailingWhitespace',
-                    '--',
-                    '4  Total',
-                    '',
-                    ''].join("\n"))
-        end
-      end
-
-      context 'when one cop is given without namespace' do
-        it 'disables the given cop' do
-          create_file('example.rb', ['if x== 0 ',
-                                     "\ty",
-                                     'end'])
-
-          cli.run(['--format', 'offenses',
-                   '--except', 'IfUnlessModifier',
-                   'example.rb'])
-          with_option = $stdout.string
-          $stdout = StringIO.new
-          cli.run(['--format', 'offenses',
-                   'example.rb'])
-          without_option = $stdout.string
-
-          expect(without_option.split($RS) - with_option.split($RS))
-            .to eq(['1  Style/IfUnlessModifier', '5  Total'])
-        end
-      end
-
-      context 'when several cops are given' do
-        %w(UnneededDisable Lint/UnneededDisable Lint).each do |cop_name|
-          it "disables the given cops including #{cop_name}" do
-            create_file('example.rb', ['if x== 100000000000000 ',
-                                       "\ty",
-                                       'end # rubocop:disable all'])
-            expect(cli.run(['--format', 'offenses',
-                            '--except',
-                            'Style/IfUnlessModifier,Style/Tab,' \
-                            "Style/SpaceAroundOperators,#{cop_name}",
-                            'example.rb'])).to eq(1)
-            expect($stderr.string).to eq('')
-            expect($stdout.string)
-              .to eq(['',
-                      '1  Style/IndentationWidth',
-                      '1  Style/NumericLiterals',
-                      '1  Style/TrailingWhitespace',
-                      '--',
-                      '3  Total',
-                      '',
-                      ''].join("\n"))
-          end
-        end
-      end
-    end
-
-    describe '--lint' do
-      it 'runs only lint cops' do
-        create_file('example.rb', ['if 0 ',
-                                   "\ty",
-                                   "\tz # rubocop:disable Style/Tab",
-                                   'end'])
-        # IfUnlessModifier depends on the configuration of LineLength.
-
-        expect(cli.run(['--format', 'simple', '--lint',
-                        'example.rb'])).to eq(1)
-        expect($stdout.string)
-          .to eq(['== example.rb ==',
-                  'W:  1:  4: Literal 0 appeared in a condition.',
-                  '',
-                  '1 file inspected, 1 offense detected',
-                  ''].join("\n"))
-      end
-    end
-
-    describe '-d/--debug' do
-      it 'shows config files' do
-        create_file('example1.rb', "\tputs 0")
-        expect(cli.run(['--debug', 'example1.rb'])).to eq(1)
-        home = File.dirname(File.dirname(File.dirname(__FILE__)))
-        expect($stdout.string.lines.grep(/configuration/).map(&:chomp))
-          .to eq(["For #{abs('')}:" \
-                  " configuration from #{home}/config/default.yml",
-                  "Inheriting configuration from #{home}/config/enabled.yml",
-                  "Inheriting configuration from #{home}/config/disabled.yml"
-                 ])
-      end
-
-      it 'shows cop names' do
-        create_file('example1.rb', 'puts 0 ')
-        file = abs('example1.rb')
-
-        expect(cli.run(['--format',
-                        'emacs',
-                        '--debug',
-                        'example1.rb'])).to eq(1)
-        expect($stdout.string.lines.to_a[-1])
-          .to eq("#{file}:1:7: C: Style/TrailingWhitespace: Trailing " \
-                 "whitespace detected.\n")
-      end
-    end
-
-    describe '-D/--display-cop-names' do
-      it 'shows cop names' do
-        create_file('example1.rb', 'puts 0 # rubocop:disable NumericLiterals ')
-        file = abs('example1.rb')
-
-        expect(cli.run(['--format', 'emacs', '--display-cop-names',
-                        'example1.rb'])).to eq(1)
-        expect($stdout.string)
-          .to eq(["#{file}:1:8: W: Lint/UnneededDisable: Unnecessary " \
-                  'disabling of `Style/NumericLiterals`.',
-                  "#{file}:1:41: C: Style/TrailingWhitespace: Trailing " \
-                  'whitespace detected.',
-                  ''].join("\n"))
-      end
-    end
-
-    describe '-E/--extra-details' do
-      it 'shows extra details' do
-        create_file('example1.rb', 'puts 0 # rubocop:disable NumericLiterals ')
-        create_file('.rubocop.yml',
-                    ['TrailingWhitespace:',
-                     '  Details: Trailing space is just sloppy.'])
-        file = abs('example1.rb')
-
-        expect(cli.run(['--format', 'emacs', '--extra-details',
-                        'example1.rb'])).to eq(1)
-        expect($stdout.string)
-          .to eq(["#{file}:1:8: W: Unnecessary disabling of " \
-                  '`Style/NumericLiterals`. ',
-                  "#{file}:1:41: C: Trailing " \
-                  'whitespace detected. Trailing space is just sloppy.',
-                  ''].join("\n"))
-
-        expect($stderr.string).to eq('')
-      end
-    end
-
-    describe '-S/--display-style-guide' do
-      it 'shows style guide entry' do
-        create_file('example1.rb', 'puts 0 ')
-        file = abs('example1.rb')
-        url =
-          'https://github.com/bbatsov/ruby-style-guide#no-trailing-whitespace'
-
-        expect(cli.run(['--format',
-                        'emacs',
-                        '--display-style-guide',
-                        'example1.rb'])).to eq(1)
-        expect($stdout.string.lines.to_a[-1])
-          .to eq("#{file}:1:7: C: Trailing whitespace detected. (#{url})\n")
-      end
-
-      it 'shows reference entry' do
-        create_file('example1.rb', '[1, 2, 3].reverse.each { |e| puts e }')
-        file = abs('example1.rb')
-        url = 'https://github.com/JuanitoFatas/fast-ruby' \
-              '#enumerablereverseeach-vs-enumerablereverse_each-code'
-
-        expect(cli.run(['--format',
-                        'emacs',
-                        '--display-style-guide',
-                        'example1.rb'])).to eq(1)
-
-        output = "#{file}:1:11: C: " \
-                 "Use `reverse_each` instead of `reverse.each`. (#{url})"
-        expect($stdout.string.lines.to_a[-1])
-          .to eq([output, ''].join("\n"))
-      end
-
-      it 'shows style guide and reference entries' do
-        create_file('example1.rb', '$foo = 1')
-        file = abs('example1.rb')
-        style_guide_link = 'https://github.com/bbatsov/ruby-style-guide' \
-                           '#instance-vars'
-        reference_link = 'http://www.zenspider.com/Languages/Ruby/QuickRef.html'
-
-        expect(cli.run(['--format',
-                        'emacs',
-                        '--display-style-guide',
-                        'example1.rb'])).to eq(1)
-
-        output = "#{file}:1:1: C: " \
-                 'Do not introduce global variables. ' \
-                 "(#{style_guide_link}, #{reference_link})"
-        expect($stdout.string.lines.to_a[-1])
-          .to eq([output, ''].join("\n"))
-      end
-    end
-
-    describe '--show-cops' do
-      shared_examples(:prints_config) do
-        it 'prints the current configuration' do
-          out = stdout.lines.to_a
-          printed_config = YAML.load(out.join)
-          cop_names = (cop_list[0] || '').split(',')
-          cop_names.each do |cop_name|
-            global_conf[cop_name].each do |key, value|
-              printed_value = printed_config[cop_name][key]
-              expect(printed_value).to eq(value)
-            end
-          end
-        end
-      end
-
-      let(:cops) { RuboCop::Cop::Cop.all }
-
-      let(:global_conf) do
-        config_path =
-          RuboCop::ConfigLoader.configuration_file_for(Dir.pwd.to_s)
-        RuboCop::ConfigLoader.configuration_from_file(config_path)
-      end
-
-      let(:stdout) { $stdout.string }
-
-      before do
-        create_file('.rubocop.yml', ['Metrics/LineLength:',
-                                     '  Max: 110'])
-        expect(cli.run(['--show-cops'] + cop_list)).to eq(0)
-      end
-
-      context 'with no args' do
-        let(:cop_list) { [] }
-
-        # Extracts the first line out of the description
-        def short_description_of_cop(cop)
-          desc = full_description_of_cop(cop)
-          desc ? desc.lines.first.strip : ''
-        end
-
-        # Gets the full description of the cop or nil if no description is set.
-        def full_description_of_cop(cop)
-          cop_config = global_conf.for_cop(cop)
-          cop_config['Description']
-        end
-
-        it 'prints all available cops and their description' do
-          cops.each do |cop|
-            expect(stdout).to include cop.cop_name
-            # Because of line breaks, we will only find the beginning.
-            expect(stdout).to include short_description_of_cop(cop)[0..60]
-          end
-        end
-
-        it 'prints all types' do
-          cops
-            .types
-            .map(&:to_s)
-            .map(&:capitalize)
-            .each { |type| expect(stdout).to include(type) }
-        end
-
-        it 'prints all cops in their right type listing' do
-          lines = stdout.lines
-          lines.slice_before(/Type /).each do |slice|
-            types = cops.types.map(&:to_s).map(&:capitalize)
-            current = types.delete(slice.shift[/Type '(?<c>[^']+)'/, 'c'])
-            # all cops in their type listing
-            cops.with_type(current).each do |cop|
-              expect(slice.any? { |l| l.include? cop.cop_name }).to be_truthy
-            end
-
-            # no cop in wrong type listing
-            types.each do |type|
-              cops.with_type(type).each do |cop|
-                expect(slice.any? { |l| l.include? cop.cop_name }).to be_falsey
-              end
-            end
-          end
-        end
-
-        include_examples :prints_config
-      end
-
-      context 'with one cop given' do
-        let(:cop_list) { ['Style/Tab'] }
-
-        it 'prints that cop and nothing else' do
-          expect(stdout).to eq(
-            ['# Supports --auto-correct',
-             'Style/Tab:',
-             '  Description: No hard tabs.',
-             '  StyleGuide: ' \
-             'https://github.com/bbatsov/ruby-style-guide#spaces-indentation',
-             '  Enabled: true',
-             '',
-             ''].join("\n"))
-        end
-
-        include_examples :prints_config
-      end
-
-      context 'with two cops given' do
-        let(:cop_list) { ['Style/Tab,Metrics/LineLength'] }
-        include_examples :prints_config
-      end
-
-      context 'with one of the cops misspelled' do
-        let(:cop_list) { ['Style/Tab,Lint/X123'] }
-
-        it 'skips the unknown cop' do
-          expect(stdout).to eq(
-            ['# Supports --auto-correct',
-             'Style/Tab:',
-             '  Description: No hard tabs.',
-             '  StyleGuide: ' \
-             'https://github.com/bbatsov/ruby-style-guide#spaces-indentation',
-             '  Enabled: true',
-             '',
-             ''].join("\n"))
-        end
-      end
-    end
-
-    describe '-f/--format' do
-      let(:target_file) { 'example.rb' }
-
-      before do
-        create_file(target_file, ['# encoding: utf-8',
-                                  '#' * 90])
-      end
-
-      describe 'builtin formatters' do
-        context 'when simple format is specified' do
-          it 'outputs with simple format' do
-            cli.run(['--format', 'simple', 'example.rb'])
-            expect($stdout.string)
-              .to include(["== #{target_file} ==",
-                           'C:  2: 81: Line is too long. [90/80]'].join("\n"))
-          end
-        end
-
-        context 'when clang format is specified' do
-          it 'outputs with clang format' do
-            create_file('example1.rb', ['# encoding: utf-8',
-                                        'x= 0 ',
-                                        '#' * 85,
-                                        'y ',
-                                        'puts x'])
-            create_file('example2.rb', ['# encoding: utf-8',
-                                        "\tx",
-                                        'def a',
-                                        '   puts',
-                                        'end'])
-            create_file('example3.rb', ['# encoding: utf-8',
-                                        'def badName',
-                                        '  if something',
-                                        '    test',
-                                        '    end',
-                                        'end'])
-            expect(cli.run(['--format', 'clang', 'example1.rb',
-                            'example2.rb', 'example3.rb']))
-              .to eq(1)
-            expect($stdout.string)
-              .to eq(['example1.rb:2:2: C: Surrounding space missing for ' \
-                      'operator =.',
-                      'x= 0 ',
-                      ' ^',
-                      'example1.rb:2:5: C: Trailing whitespace detected.',
-                      'x= 0 ',
-                      '    ^',
-                      'example1.rb:3:81: C: Line is too long. [85/80]',
-                      '###################################################' \
-                      '##################################',
-                      '                                                   ' \
-                      '                             ^^^^^',
-                      'example1.rb:4:2: C: Trailing whitespace detected.',
-                      'y ',
-                      ' ^',
-                      'example2.rb:1:1: C: Incorrect indentation detected' \
-                      ' (column 0 instead of 1).',
-                      '# encoding: utf-8',
-                      '^^^^^^^^^^^^^^^^^',
-                      'example2.rb:2:1: C: Tab detected.',
-                      "\tx",
-                      '^',
-                      'example2.rb:2:2: C: Indentation of first line in file ' \
-                      'detected.',
-                      '	x',
-                      ' ^',
-                      'example2.rb:3:1: C: Inconsistent indentation ' \
-                      'detected.',
-                      'def a',
-                      '^^^^^',
-                      'example2.rb:4:1: C: Use 2 (not 3) spaces for ' \
-                      'indentation.',
-                      '   puts',
-                      '^^^',
-                      'example3.rb:2:5: C: Use snake_case for method names.',
-                      'def badName',
-                      '    ^^^^^^^',
-                      'example3.rb:3:3: C: Use a guard clause instead of ' \
-                      'wrapping the code inside a conditional expression.',
-                      '  if something',
-                      '  ^^',
-                      'example3.rb:3:3: C: Favor modifier if usage ' \
-                      'when having a single-line body. Another good ' \
-                      'alternative is the usage of control flow &&/||.',
-                      '  if something',
-                      '  ^^',
-                      'example3.rb:5:5: W: end at 5, 4 is not aligned ' \
-                      'with if at 3, 2.',
-                      '    end',
-                      '    ^^^',
-                      '',
-                      '3 files inspected, 13 offenses detected',
-                      ''].join("\n"))
-          end
-        end
-
-        context 'when emacs format is specified' do
-          it 'outputs with emacs format' do
-            create_file('example1.rb', ['# encoding: utf-8',
-                                        'x= 0 ',
-                                        'y ',
-                                        'puts x'])
-            create_file('example2.rb', ['# encoding: utf-8',
-                                        "\tx = 0",
-                                        'puts x'])
-            expect(cli.run(['--format', 'emacs', 'example1.rb',
-                            'example2.rb'])).to eq(1)
-            expected_output =
-              ["#{abs('example1.rb')}:2:2: C: Surrounding space missing" \
-               ' for operator `=`.',
-               "#{abs('example1.rb')}:2:5: C: Trailing whitespace detected.",
-               "#{abs('example1.rb')}:3:2: C: Trailing whitespace detected.",
-               "#{abs('example2.rb')}:1:1: C: Incorrect indentation detected" \
-               ' (column 0 instead of 1).',
-               "#{abs('example2.rb')}:2:1: C: Tab detected.",
-               "#{abs('example2.rb')}:2:2: C: Indentation of first line in " \
-               'file detected.',
-               "#{abs('example2.rb')}:3:1: C: Inconsistent indentation " \
-               'detected.',
-               ''].join("\n")
-            expect($stdout.string).to eq(expected_output)
-          end
-        end
-
-        context 'when unknown format name is specified' do
-          it 'aborts with error message' do
-            expect(cli.run(['--format', 'unknown', 'example.rb'])).to eq(1)
-            expect($stderr.string)
-              .to include('No formatter for "unknown"')
-          end
-        end
-
-        context 'when ambiguous format name is specified' do
-          it 'aborts with error message' do
-            # Both 'files' and 'fuubar' start with an 'f'.
-            expect(cli.run(['--format', 'f', 'example.rb'])).to eq(1)
-            expect($stderr.string)
-              .to include('Cannot determine formatter for "f"')
-          end
-        end
-      end
-
-      describe 'custom formatter' do
-        let(:target_file) { abs('example.rb') }
-
-        context 'when a class name is specified' do
-          it 'uses the class as a formatter' do
-            module MyTool
-              class RuboCopFormatter < RuboCop::Formatter::BaseFormatter
-                def started(all_files)
-                  output.puts "started: #{all_files.join(',')}"
-                end
-
-                def file_started(file, _options)
-                  output.puts "file_started: #{file}"
-                end
-
-                def file_finished(file, _offenses)
-                  output.puts "file_finished: #{file}"
-                end
-
-                def finished(processed_files)
-                  output.puts "finished: #{processed_files.join(',')}"
-                end
-              end
-            end
-
-            cli.run(['--format', 'MyTool::RuboCopFormatter', 'example.rb'])
-            expect($stdout.string).to eq(["started: #{target_file}",
-                                          "file_started: #{target_file}",
-                                          "file_finished: #{target_file}",
-                                          "finished: #{target_file}",
-                                          ''].join("\n"))
-          end
-        end
-
-        context 'when unknown class name is specified' do
-          it 'aborts with error message' do
-            args = '--format UnknownFormatter example.rb'
-            expect(cli.run(args.split)).to eq(1)
-            expect($stderr.string).to include('UnknownFormatter')
-          end
-        end
-      end
-
-      it 'can be used multiple times' do
-        cli.run(['--format', 'simple', '--format', 'emacs', 'example.rb'])
-        expect($stdout.string)
-          .to include(["== #{target_file} ==",
-                       'C:  2: 81: Line is too long. [90/80]',
-                       "#{abs(target_file)}:2:81: C: Line is too long. " \
-                       '[90/80]'].join("\n"))
-      end
-    end
-
-    describe '-o/--out option' do
-      let(:target_file) { 'example.rb' }
-
-      before do
-        create_file(target_file, ['# encoding: utf-8',
-                                  '#' * 90])
-      end
-
-      it 'redirects output to the specified file' do
-        cli.run(['--out', 'output.txt', target_file])
-        expect(File.read('output.txt')).to include('Line is too long.')
-      end
-
-      it 'is applied to the previously specified formatter' do
-        cli.run(['--format', 'simple',
-                 '--format', 'emacs', '--out', 'emacs_output.txt',
-                 target_file])
-
-        expect($stdout.string).to eq(["== #{target_file} ==",
-                                      'C:  2: 81: Line is too long. [90/80]',
-                                      '',
-                                      '1 file inspected, 1 offense detected',
-                                      ''].join("\n"))
-
-        expect(File.read('emacs_output.txt'))
-          .to eq(["#{abs(target_file)}:2:81: C: Line is too long. [90/80]",
-                  ''].join("\n"))
-      end
-    end
-
-    describe '--fail-level option' do
-      let(:target_file) { 'example.rb' }
-
-      before do
-        create_file(target_file, ['# encoding: utf-8',
-                                  'def f',
-                                  ' x',
-                                  'end'])
-      end
-
-      after do
-        expect($stderr.string).to eq('')
-        expect($stdout.string)
-          .to include('1 file inspected, 1 offense detected')
-      end
-
-      it 'fails when option is less than the severity level' do
-        expect(cli.run(['--fail-level', 'refactor', target_file])).to eq(1)
-        expect(cli.run(['--fail-level', 'autocorrect', target_file])).to eq(1)
-      end
-
-      it 'fails when option is equal to the severity level' do
-        expect(cli.run(['--fail-level', 'convention', target_file])).to eq(1)
-      end
-
-      it 'succeeds when option is greater than the severity level' do
-        expect(cli.run(['--fail-level', 'warning', target_file])).to eq(0)
-      end
-
-      context 'with --auto-correct' do
-        after do
-          expect($stdout.string.lines.to_a.last)
-            .to eq('1 file inspected, 1 offense detected, 1 offense corrected' \
-                   "\n")
-        end
-
-        it 'fails when option is autocorrect and all offenses are ' \
-           'autocorrected' do
-          expect(cli.run(['--auto-correct', '--format', 'simple',
-                          '--fail-level', 'autocorrect',
-                          target_file])).to eq(1)
-        end
-
-        it 'fails when option is A and all offenses are autocorrected' do
-          expect(cli.run(['--auto-correct', '--format', 'simple',
-                          '--fail-level', 'A',
-                          target_file])).to eq(1)
-        end
-
-        it 'succeeds when option is not given and all offenses are ' \
-           'autocorrected' do
-          expect(cli.run(['--auto-correct', '--format', 'simple',
-                          target_file])).to eq(0)
-        end
-
-        it 'succeeds when option is refactor and all offenses are ' \
-           'autocorrected' do
-          expect(cli.run(['--auto-correct', '--format', 'simple',
-                          '--fail-level', 'refactor',
-                          target_file])).to eq(0)
-        end
-      end
-    end
-
-    describe 'with --auto-correct and disabled offense' do
-      let(:target_file) { 'example.rb' }
-      after do
-        expect($stdout.string.lines.to_a.last)
-          .to eq('1 file inspected, no offenses detected' \
-                 "\n")
-      end
-      it 'succeeds when there is only a disabled offense' do
-        create_file(target_file, ['# encoding: utf-8',
-                                  'def f',
-                                  ' x # rubocop:disable Style/IndentationWidth',
-                                  'end'])
-
-        expect(cli.run(['--auto-correct', '--format', 'simple',
-                        '--fail-level', 'autocorrect',
-                        target_file])).to eq(0)
-      end
-    end
-
-    describe '--force-exclusion' do
-      let(:target_file) { 'example.rb' }
-
-      before do
-        create_file(target_file, ['# encoding: utf-8',
-                                  '#' * 90])
-
-        create_file('.rubocop.yml', ['AllCops:',
-                                     '  Exclude:',
-                                     "    - #{target_file}"])
-      end
-
-      it 'excludes files specified in the configuration Exclude ' \
-         'even if they are explicitly passed as arguments' do
-        expect(cli.run(['--force-exclusion', target_file])).to eq(0)
-      end
-    end
-
-    describe '--stdin' do
-      it 'causes source code to be read from stdin' do
-        begin
-          $stdin = StringIO.new('p $/')
-          argv   = ['--only=Style/SpecialGlobalVars',
-                    '--format=simple',
-                    '--stdin',
-                    'fake.rb']
-          expect(cli.run(argv)).to eq(1)
-          expect($stdout.string).to eq([
-            '== fake.rb ==',
-            'C:  1:  3: Prefer $INPUT_RECORD_SEPARATOR or $RS from the ' \
-            "stdlib 'English' module over $/.",
-            '',
-            '1 file inspected, 1 offense detected',
-            ''].join("\n"))
-        ensure
-          $stdin = STDIN
-        end
-      end
-
-      it 'requires a file path' do
-        begin
-          $stdin = StringIO.new('p $/')
-          argv   = ['--only=Style/SpecialGlobalVars',
-                    '--format=simple',
-                    '--stdin']
-          expect(cli.run(argv)).to eq(1)
-          expect($stderr.string).to include(
-            '-s/--stdin requires exactly one path.')
-        ensure
-          $stdin = STDIN
-        end
-      end
-
-      it 'does not accept more than one file path' do
-        begin
-          $stdin = StringIO.new('p $/')
-          argv   = ['--only=Style/SpecialGlobalVars',
-                    '--format=simple',
-                    '--stdin',
-                    'fake1.rb',
-                    'fake2.rb']
-          expect(cli.run(argv)).to eq(1)
-          expect($stderr.string).to include(
-            '-s/--stdin requires exactly one path.')
-        ensure
-          $stdin = STDIN
-        end
-      end
-
-      it 'prints corrected code to stdout if --autocorrect is used' do
-        begin
-          $stdin = StringIO.new('p $/')
-          argv   = ['--auto-correct',
-                    '--only=Style/SpecialGlobalVars',
-                    '--format=simple',
-                    '--stdin',
-                    'fake.rb']
-          expect(cli.run(argv)).to eq(0)
-          expect($stdout.string).to eq([
-            '== fake.rb ==',
-            'C:  1:  3: [Corrected] Prefer $INPUT_RECORD_SEPARATOR or $RS ' \
-            "from the stdlib 'English' module over $/.",
-            '',
-            '1 file inspected, 1 offense detected, 1 offense corrected',
-            '====================',
-            'p $INPUT_RECORD_SEPARATOR'].join("\n"))
-        ensure
-          $stdin = STDIN
-        end
-      end
-    end
-  end
-
   context 'when interrupted' do
     it 'returns 1' do
       allow_any_instance_of(RuboCop::Runner)
